FFMPEG: let nuttx -W cflags valid

+++ libavcodec/dnxhdenc.h
@@ -28,13 +28,11 @@
 
 #include "config.h"
 
-#include "libavutil/mem_internal.h"
-
 #include "mpegvideo.h"
 #include "dnxhddata.h"
 
 typedef struct RCCMPEntry {
-    uint32_t mb;
+    uint16_t mb;
     int value;
 } RCCMPEntry;
 
@@ -100,7 +98,7 @@ typedef struct DNXHDEncContext {
     unsigned qscale;
     unsigned lambda;
 
-    uint32_t *mb_bits;
+    uint16_t *mb_bits;
     uint8_t  *mb_qscale;
 
     RCCMPEntry *mb_cmp;
